Switch to using TLS certs for docker validation instead of tunneling via rdocker
secrets/scripts/validate_docker_certs.sh will do some basic validation
on the existing certs in a directory and make sure they match the
expected environment

secrets/scripts/create_docker_certs.sh will create TLS server and client
certs for docker

bootstrap/make-docker-certs.sh will use the two above scripts to valdiate
and (if necessary) create TLS certs for an environment

bootstrap/ansible/configure-docker-certs.yaml will upload the certs to
the appropraite droplet and reconfigure docker to use them

The existing ansible-inventory script connects to docker to create volumes.
This previously worked with rdocker but now that we need to use ansible
to make it possible to connect to docker, we have a bit of a chicken-and-egg
problem.

So there's no a staging-inventory-no-volumes.sh inventory script that will
let ansible connect but won't provide any volume variables to the playbook.

So it's now important to use the correct inventory script for a given playbook.

The readme is now updatd to include all of the above in the right order.