Move compose files to a compose directory for sane grouping
While testing this, I discovered that the containers in docker are labeled with a hash of the config that was used to create them -- which includes the path that was used to find the compose file!

So `cd dir && compose ./file.yaml`  is different from `compose ./dir/file.yaml` and causes conflicts.

So I needed to tear down my old containers and rebuild them from a new directory structure.

So I also added the ability to specify "down" to the compose-environment.sh script.

I also made this script work with the new docker cert setup, which I'd neglected to do in my last commit. Oops.